fix: refactored slightly
