--- src/routes/ResourceSection/ResourceSectionComponents/PageTitle.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const PageTitle = () => {
-  return (
-    <div className="PageTitle">
-      <span className="text">resources</span>
-      <span className="arrow">&gt; &gt; &gt;</span>
-    </div>
-  );
-};
-
-export default PageTitle;
