--- src/routes/ResourceSection/ResourceSectionComponents/ResourcesBlurb.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const ResourcesBlurb = () => {
-  return (
-    <div className="ResourcesBlurb">See our resources from previous events and more!</div>
-  );
-};
-
-export default ResourcesBlurb;
